test: move C-binding unit tests to ffi-bench, drop zstd dev-dependency
The library crate now links zero C bindings. Every test that cross-checks
our output against libzstd lives in the sibling ffi-bench crate:

- frame compressor FFI roundtrips (+ fuzz-artifact interop replay)
- streaming encoder FFI roundtrips
- huff0 weight-description / 4-stream / level-22 literal-section conformance
- match-generator level-22 sequence-stream conformance
- dictionary roundtrip through the C decoder
- corpus roundtrip through the C decoder

White-box cases keep their pure-Rust capture in the library behind a
testing facade (bench_internals) and feed only the result to the C side
in ffi-bench. The our-decoder corpus roundtrips stay in the library.
Removed the zstd dev-dependency from the library crate; fixed the moved
integration tests to resolve the corpus fixture from either crate root.
