+++ zstd/src/encoding/frame_compressor.rs
@@ -2312,9 +2312,8 @@ mod tests {
     use alloc::vec;
 
     use super::FrameCompressor;
-    use crate::blocks::block::BlockType;
     use crate::common::{MAGIC_NUM, MAX_BLOCK_SIZE};
-    use crate::decoding::{FrameDecoder, block_decoder, frame::read_frame_header};
+    use crate::decoding::FrameDecoder;
     use crate::encoding::{Matcher, Sequence};
     use alloc::vec::Vec;
 
@@ -2330,461 +2329,9 @@ mod tests {
         data
     }
 
-    fn first_block_type(frame: &[u8]) -> BlockType {
-        let (_, header_size) = read_frame_header(frame).expect("frame header should parse");
-        let mut decoder = block_decoder::new();
-        let (header, _) = decoder
-            .read_block_header(&frame[header_size as usize..])
-            .expect("block header should parse");
-        header.block_type
-    }
-
-    /// Frame content size is written correctly and C zstd can decompress the output.
-    #[cfg(feature = "std")]
-    #[test]
-    fn fcs_header_written_and_c_zstd_compatible() {
-        let levels = [
-            crate::encoding::CompressionLevel::Uncompressed,
-            crate::encoding::CompressionLevel::Fastest,
-            crate::encoding::CompressionLevel::Default,
-            crate::encoding::CompressionLevel::Better,
-            crate::encoding::CompressionLevel::Best,
-        ];
-        let fcs_2byte = vec![0xCDu8; 300]; // 300 bytes → 2-byte FCS (256..=65791 range)
-        let large = vec![0xABu8; 100_000];
-        let inputs: [&[u8]; 5] = [
-            &[],
-            &[0x00],
-            b"abcdefghijklmnopqrstuvwxy\n",
-            &fcs_2byte,
-            &large,
-        ];
-        for level in levels {
-            for data in &inputs {
-                let compressed = crate::encoding::compress_to_vec(*data, level);
-                // Verify FCS is present and correct
-                let header = crate::decoding::frame::read_frame_header(compressed.as_slice())
-                    .unwrap()
-                    .0;
-                assert_eq!(
-                    header.frame_content_size(),
-                    data.len() as u64,
-                    "FCS mismatch for len={} level={:?}",
-                    data.len(),
-                    level,
-                );
-                // Confirm the FCS field is actually present in the header
-                // (not just the decoder returning 0 for absent FCS).
-                assert_ne!(
-                    header.descriptor.frame_content_size_bytes().unwrap(),
-                    0,
-                    "FCS field must be present for len={} level={:?}",
-                    data.len(),
-                    level,
-                );
-                // Verify C zstd can decompress
-                let mut decoded = Vec::new();
-                zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap_or_else(
-                    |e| {
-                        panic!(
-                            "C zstd decode failed for len={} level={level:?}: {e}",
-                            data.len()
-                        )
-                    },
-                );
-                assert_eq!(
-                    decoded.as_slice(),
-                    *data,
-                    "C zstd roundtrip failed for len={}",
-                    data.len()
-                );
-            }
-        }
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn source_size_hint_fastest_remains_ffi_compatible_small_input() {
-        let data = vec![0xAB; 2047];
-        let compressed = {
-            let mut compressor = FrameCompressor::new(super::CompressionLevel::Fastest);
-            compressor.set_source_size_hint(data.len() as u64);
-            compressor.set_source(data.as_slice());
-            let mut out = Vec::new();
-            compressor.set_drain(&mut out);
-            compressor.compress();
-            out
-        };
-
-        let mut decoded = Vec::new();
-        zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap();
-        assert_eq!(decoded, data);
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn small_hinted_default_frame_uses_single_segment_header() {
-        let data = generate_data(0xD15E_A5ED, 1024);
-        let compressed = {
-            let mut compressor = FrameCompressor::new(super::CompressionLevel::Default);
-            compressor.set_source_size_hint(data.len() as u64);
-            compressor.set_source(data.as_slice());
-            let mut out = Vec::new();
-            compressor.set_drain(&mut out);
-            compressor.compress();
-            out
-        };
-
-        let (frame_header, _) = read_frame_header(compressed.as_slice()).unwrap();
-        assert!(
-            frame_header.descriptor.single_segment_flag(),
-            "small hinted default frames should use single-segment header for Rust/FFI parity"
-        );
-        assert_eq!(frame_header.frame_content_size(), data.len() as u64);
-        let mut decoded = Vec::new();
-        zstd::stream::copy_decode(compressed.as_slice(), &mut decoded)
-            .expect("ffi decoder must accept single-segment small hinted default frame");
-        assert_eq!(decoded, data);
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn small_hinted_numeric_default_levels_use_single_segment_header() {
-        let data = generate_data(0xA11C_E003, 1024);
-        for level in [
-            super::CompressionLevel::Level(0),
-            super::CompressionLevel::Level(3),
-        ] {
-            let compressed = {
-                let mut compressor = FrameCompressor::new(level);
-                compressor.set_source_size_hint(data.len() as u64);
-                compressor.set_source(data.as_slice());
-                let mut out = Vec::new();
-                compressor.set_drain(&mut out);
-                compressor.compress();
-                out
-            };
-
-            let (frame_header, _) = read_frame_header(compressed.as_slice()).unwrap();
-            assert!(
-                frame_header.descriptor.single_segment_flag(),
-                "small hinted numeric default level frames should use single-segment header (level={level:?})"
-            );
-            assert_eq!(frame_header.frame_content_size(), data.len() as u64);
-            let mut decoded = Vec::new();
-            zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap_or_else(|e| {
-                panic!(
-                    "ffi decoder must accept single-segment small hinted numeric default level frame (level={level:?}): {e}"
-                )
-            });
-            assert_eq!(decoded, data);
-        }
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn source_size_hint_levels_remain_ffi_compatible_small_inputs_matrix() {
-        let levels = [
-            super::CompressionLevel::Fastest,
-            super::CompressionLevel::Default,
-            super::CompressionLevel::Better,
-            super::CompressionLevel::Best,
-            super::CompressionLevel::Level(-1),
-            super::CompressionLevel::Level(2),
-            super::CompressionLevel::Level(3),
-            super::CompressionLevel::Level(4),
-            super::CompressionLevel::Level(11),
-        ];
-        let sizes = [
-            511usize, 512, 513, 1023, 1024, 1536, 2047, 2048, 4095, 4096, 8191, 16_384, 16_385,
-        ];
-
-        for (seed_idx, seed) in [11u64, 23, 41].into_iter().enumerate() {
-            for &size in &sizes {
-                let data = generate_data(seed + seed_idx as u64, size);
-                for &level in &levels {
-                    let compressed = {
-                        let mut compressor = FrameCompressor::new(level);
-                        compressor.set_source_size_hint(data.len() as u64);
-                        compressor.set_source(data.as_slice());
-                        let mut out = Vec::new();
-                        compressor.set_drain(&mut out);
-                        compressor.compress();
-                        out
-                    };
-                    if matches!(size, 511 | 512) {
-                        let (frame_header, _) = read_frame_header(compressed.as_slice()).unwrap();
-                        assert_eq!(
-                            frame_header.descriptor.single_segment_flag(),
-                            size == 512,
-                            "single_segment 511/512 boundary mismatch: level={level:?} size={size}"
-                        );
-                    }
-
-                    let mut decoded = Vec::new();
-                    zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap_or_else(
-                        |e| {
-                            panic!(
-                                "ffi decode failed with source-size hint: level={level:?} size={size} seed={} err={e}",
-                                seed + seed_idx as u64
-                            )
-                        },
-                    );
-                    assert_eq!(
-                        decoded,
-                        data,
-                        "hinted ffi roundtrip mismatch: level={level:?} size={size} seed={}",
-                        seed + seed_idx as u64
-                    );
-                }
-            }
-        }
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn hinted_levels_use_single_segment_header_symmetrically() {
-        let levels = [
-            super::CompressionLevel::Fastest,
-            super::CompressionLevel::Default,
-            super::CompressionLevel::Better,
-            super::CompressionLevel::Best,
-            super::CompressionLevel::Level(0),
-            super::CompressionLevel::Level(2),
-            super::CompressionLevel::Level(3),
-            super::CompressionLevel::Level(4),
-            super::CompressionLevel::Level(11),
-        ];
-        for (seed_idx, seed) in [7u64, 23, 41].into_iter().enumerate() {
-            let size = 1024 + seed_idx * 97;
-            let data = generate_data(seed, size);
-            for &level in &levels {
-                let compressed = {
-                    let mut compressor = FrameCompressor::new(level);
-                    compressor.set_source_size_hint(data.len() as u64);
-                    compressor.set_source(data.as_slice());
-                    let mut out = Vec::new();
-                    compressor.set_drain(&mut out);
-                    compressor.compress();
-                    out
-                };
-                let (frame_header, _) = read_frame_header(compressed.as_slice()).unwrap();
-                assert!(
-                    frame_header.descriptor.single_segment_flag(),
-                    "hinted frame should be single-segment for level={level:?} size={}",
-                    data.len()
-                );
-                assert_eq!(frame_header.frame_content_size(), data.len() as u64);
-                let mut decoded = Vec::new();
-                zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap_or_else(|e| {
-                    panic!(
-                        "ffi decode failed for hinted single-segment parity: level={level:?} size={} err={e}",
-                        data.len()
-                    )
-                });
-                assert_eq!(decoded, data);
-            }
-        }
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn hinted_levels_pin_511_512_single_segment_boundary() {
-        let levels = [
-            super::CompressionLevel::Fastest,
-            super::CompressionLevel::Default,
-            super::CompressionLevel::Better,
-            super::CompressionLevel::Best,
-            super::CompressionLevel::Level(0),
-            super::CompressionLevel::Level(2),
-            super::CompressionLevel::Level(3),
-            super::CompressionLevel::Level(4),
-            super::CompressionLevel::Level(11),
-        ];
-        for (seed_idx, seed) in [7u64, 23, 41].into_iter().enumerate() {
-            for &size in &[511usize, 512] {
-                let data = generate_data(seed + seed_idx as u64, size);
-                for &level in &levels {
-                    let compressed = {
-                        let mut compressor = FrameCompressor::new(level);
-                        compressor.set_source_size_hint(data.len() as u64);
-                        compressor.set_source(data.as_slice());
-                        let mut out = Vec::new();
-                        compressor.set_drain(&mut out);
-                        compressor.compress();
-                        out
-                    };
-                    let (frame_header, _) = read_frame_header(compressed.as_slice()).unwrap();
-                    assert_eq!(
-                        frame_header.descriptor.single_segment_flag(),
-                        size == 512,
-                        "single_segment 511/512 boundary mismatch: level={level:?} size={size}"
-                    );
-                    let mut decoded = Vec::new();
-                    zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap_or_else(
-                        |e| {
-                            panic!(
-                                "ffi decode failed at single-segment boundary: level={level:?} size={size} seed={} err={e}",
-                                seed + seed_idx as u64
-                            )
-                        },
-                    );
-                    assert_eq!(decoded, data);
-                }
-            }
-        }
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn fastest_random_block_uses_raw_fast_path() {
-        let data = generate_data(0xC0FF_EE11, 10 * 1024);
-        let compressed =
-            crate::encoding::compress_to_vec(data.as_slice(), super::CompressionLevel::Fastest);
-
-        assert_eq!(first_block_type(&compressed), BlockType::Raw);
-
-        let mut decoded = Vec::new();
-        zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap();
-        assert_eq!(decoded, data);
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn default_random_block_uses_raw_fast_path() {
-        let data = generate_data(0xD15E_A5ED, 10 * 1024);
-        let compressed =
-            crate::encoding::compress_to_vec(data.as_slice(), super::CompressionLevel::Default);
-
-        assert_eq!(first_block_type(&compressed), BlockType::Raw);
-
-        let mut decoded = Vec::new();
-        zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap();
-        assert_eq!(decoded, data);
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn best_random_block_uses_raw_fast_path() {
-        let data = generate_data(0xB35C_AFE1, 10 * 1024);
-        let compressed =
-            crate::encoding::compress_to_vec(data.as_slice(), super::CompressionLevel::Best);
-
-        assert_eq!(first_block_type(&compressed), BlockType::Raw);
-
-        let mut decoded = Vec::new();
-        zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap();
-        assert_eq!(decoded, data);
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn level2_random_block_uses_raw_fast_path() {
-        let data = generate_data(0xA11C_E222, 10 * 1024);
-        let compressed =
-            crate::encoding::compress_to_vec(data.as_slice(), super::CompressionLevel::Level(2));
-
-        assert_eq!(first_block_type(&compressed), BlockType::Raw);
-
-        let mut decoded = Vec::new();
-        zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap();
-        assert_eq!(decoded, data);
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn better_random_block_uses_raw_fast_path() {
-        let data = generate_data(0xBE77_E111, 10 * 1024);
-        let compressed =
-            crate::encoding::compress_to_vec(data.as_slice(), super::CompressionLevel::Better);
-
-        assert_eq!(first_block_type(&compressed), BlockType::Raw);
-
-        let mut decoded = Vec::new();
-        zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap();
-        assert_eq!(decoded, data);
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn compressible_logs_do_not_fall_back_to_raw_fast_path() {
-        let mut data = Vec::with_capacity(16 * 1024);
-        const LINE: &[u8] =
-            b"ts=2026-04-10T00:00:00Z level=INFO tenant=demo op=flush table=orders\n";
-        while data.len() < 16 * 1024 {
-            let remaining = 16 * 1024 - data.len();
-            data.extend_from_slice(&LINE[..LINE.len().min(remaining)]);
-        }
-
-        fn assert_not_raw_for_level(data: &[u8], level: super::CompressionLevel) {
-            let compressed = crate::encoding::compress_to_vec(data, level);
-            assert_ne!(first_block_type(&compressed), BlockType::Raw);
-            assert!(
-                compressed.len() < data.len(),
-                "compressible input should remain compressible for level={level:?}"
-            );
-            let mut decoded = Vec::new();
-            zstd::stream::copy_decode(compressed.as_slice(), &mut decoded).unwrap();
-            assert_eq!(decoded, data);
-        }
-
-        assert_not_raw_for_level(data.as_slice(), super::CompressionLevel::Fastest);
-        assert_not_raw_for_level(data.as_slice(), super::CompressionLevel::Default);
-        assert_not_raw_for_level(data.as_slice(), super::CompressionLevel::Level(3));
-        assert_not_raw_for_level(data.as_slice(), super::CompressionLevel::Better);
-        assert_not_raw_for_level(data.as_slice(), super::CompressionLevel::Best);
-    }
-
-    #[cfg(feature = "std")]
-    #[test]
-    fn hinted_small_compressible_frames_use_single_segment_across_levels() {
-        let mut data = Vec::with_capacity(4 * 1024);
-        const LINE: &[u8] =
-            b"ts=2026-04-10T00:00:00Z level=INFO tenant=demo op=flush table=orders\n";
-        while data.len() < 4 * 1024 {
-            let remaining = 4 * 1024 - data.len();
-            data.extend_from_slice(&LINE[..LINE.len().min(remaining)]);
-        }
-
-        for level in [
-            super::CompressionLevel::Fastest,
-            super::CompressionLevel::Default,
-            super::CompressionLevel::Better,
-            super::CompressionLevel::Best,
-            super::CompressionLevel::Level(0),
-            super::CompressionLevel::Level(3),
-            super::CompressionLevel::Level(4),
-            super::CompressionLevel::Level(11),
-        ] {
-            let compressed = {
-                let mut compressor = FrameCompressor::new(level);
-                compressor.set_source_size_hint(data.len() as u64);
-                compressor.set_source(data.as_slice());
-                let mut out = Vec::new();
-                compressor.set_drain(&mut out);
-                compressor.compress();
-                out
-            };
-            let (frame_header, _) = read_frame_header(compressed.as_slice()).unwrap();
-            assert!(
-                frame_header.descriptor.single_segment_flag(),
-                "hinted small compressible frame should use single-segment (level={level:?})"
-            );
-            assert_ne!(
-                first_block_type(&compressed),
-                BlockType::Raw,
-                "compressible hinted frame should stay off raw fast path (level={level:?})"
-            );
-            assert!(
-                compressed.len() < data.len(),
-                "compressible hinted frame should still shrink (level={level:?})"
-            );
-            let mut decoded = Vec::new();
-            zstd::stream::copy_decode(compressed.as_slice(), &mut decoded)
-                .unwrap_or_else(|e| panic!("ffi decode failed (level={level:?}): {e}"));
-            assert_eq!(decoded, data);
-        }
-    }
+    // Cross-implementation parity tests (compress here, decode through the C
+    // bindings) moved to `ffi-bench/tests/frame_compressor_ffi.rs` so the
+    // library crate never links libzstd.
 
     struct NoDictionaryMatcher {
         last_space: Vec<u8>,
@@ -2871,10 +2418,6 @@ mod tests {
         let mut decoded = Vec::with_capacity(mock_data.len());
         decoder.decode_all_to_vec(&output, &mut decoded).unwrap();
         assert_eq!(mock_data, decoded);
-
-        let mut decoded = Vec::new();
-        zstd::stream::copy_decode(output.as_slice(), &mut decoded).unwrap();
-        assert_eq!(mock_data, decoded);
     }
 
     #[test]
@@ -2906,10 +2449,6 @@ mod tests {
         let mut decoder = FrameDecoder::new();
         let mut decoded = Vec::with_capacity(mock_data.len());
         decoder.decode_all_to_vec(&output, &mut decoded).unwrap();
-        assert_eq!(mock_data, decoded);
-
-        let mut decoded = Vec::new();
-        zstd::stream::copy_decode(output.as_slice(), &mut decoded).unwrap();
         assert_eq!(mock_data, decoded);
     }
 
@@ -2967,14 +2506,6 @@ mod tests {
         let mut decoded = Vec::with_capacity(data.len());
         decoder.decode_all_to_vec(&with_dict, &mut decoded).unwrap();
         assert_eq!(decoded, data);
-
-        let mut ffi_decoder = zstd::bulk::Decompressor::with_dictionary(dict_raw).unwrap();
-        let mut ffi_decoded = Vec::with_capacity(data.len());
-        let ffi_written = ffi_decoder
-            .decompress_to_buffer(with_dict.as_slice(), &mut ffi_decoded)
-            .unwrap();
-        assert_eq!(ffi_written, data.len());
-        assert_eq!(ffi_decoded, data);
     }
 
     #[cfg(all(feature = "dict_builder", feature = "std"))]
@@ -3245,16 +2776,6 @@ mod tests {
             .decode_all_to_vec(&frame, &mut decoded)
             .expect("single-segment dict frame must decode");
         assert_eq!(decoded, payload);
-
-        // Reference-decoder cross-check: our serializer and parser could
-        // share a bug and still self-roundtrip, so the on-wire
-        // single-segment + dictionary-ID layout must also decode through
-        // the C implementation.
-        let ffi_decoded = zstd::bulk::Decompressor::with_dictionary(dict_raw)
-            .expect("reference decompressor must accept the dictionary")
-            .decompress(&frame, payload.len() + 64)
-            .expect("reference zstd must accept the single-segment dict frame");
-        assert_eq!(ffi_decoded, payload);
     }
 
     // Regression test: `heap_size()` must count the retained Huffman tables
@@ -4681,101 +4202,9 @@ mod tests {
         );
     }
 
-    #[cfg(feature = "std")]
-    #[test]
-    fn fuzz_targets() {
-        use std::io::Read;
-        fn decode_szstd(data: &mut dyn std::io::Read) -> Vec<u8> {
-            let mut decoder = crate::decoding::StreamingDecoder::new(data).unwrap();
-            let mut result: Vec<u8> = Vec::new();
-            decoder.read_to_end(&mut result).expect("Decoding failed");
-            result
-        }
-
-        fn decode_szstd_writer(mut data: impl Read) -> Vec<u8> {
-            let mut decoder = crate::decoding::FrameDecoder::new();
-            decoder.reset(&mut data).unwrap();
-            let mut result = vec![];
-            while !decoder.is_finished() || decoder.can_collect() > 0 {
-                decoder
-                    .decode_blocks(
-                        &mut data,
-                        crate::decoding::BlockDecodingStrategy::UptoBytes(1024 * 1024),
-                    )
-                    .unwrap();
-                decoder.collect_to_writer(&mut result).unwrap();
-            }
-            result
-        }
-
-        fn encode_zstd(data: &[u8]) -> Result<Vec<u8>, std::io::Error> {
-            zstd::stream::encode_all(std::io::Cursor::new(data), 3)
-        }
-
-        fn encode_szstd_uncompressed(data: &mut dyn std::io::Read) -> Vec<u8> {
-            let mut input = Vec::new();
-            data.read_to_end(&mut input).unwrap();
-
-            crate::encoding::compress_to_vec(
-                input.as_slice(),
-                crate::encoding::CompressionLevel::Uncompressed,
-            )
-        }
-
-        fn encode_szstd_compressed(data: &mut dyn std::io::Read) -> Vec<u8> {
-            let mut input = Vec::new();
-            data.read_to_end(&mut input).unwrap();
-
-            crate::encoding::compress_to_vec(
-                input.as_slice(),
-                crate::encoding::CompressionLevel::Fastest,
-            )
-        }
-
-        fn decode_zstd(data: &[u8]) -> Result<Vec<u8>, std::io::Error> {
-            let mut output = Vec::new();
-            zstd::stream::copy_decode(data, &mut output)?;
-            Ok(output)
-        }
-        if std::fs::exists("fuzz/artifacts/interop").unwrap_or(false) {
-            for file in std::fs::read_dir("fuzz/artifacts/interop").unwrap() {
-                if file.as_ref().unwrap().file_type().unwrap().is_file() {
-                    let data = std::fs::read(file.unwrap().path()).unwrap();
-                    let data = data.as_slice();
-                    // Decoding
-                    let compressed = encode_zstd(data).unwrap();
-                    let decoded = decode_szstd(&mut compressed.as_slice());
-                    let decoded2 = decode_szstd_writer(&mut compressed.as_slice());
-                    assert!(
-                        decoded == data,
-                        "Decoded data did not match the original input during decompression"
-                    );
-                    assert_eq!(
-                        decoded2, data,
-                        "Decoded data did not match the original input during decompression"
-                    );
-
-                    // Encoding
-                    // Uncompressed encoding
-                    let mut input = data;
-                    let compressed = encode_szstd_uncompressed(&mut input);
-                    let decoded = decode_zstd(&compressed).unwrap();
-                    assert_eq!(
-                        decoded, data,
-                        "Decoded data did not match the original input during compression"
-                    );
-                    // Compressed encoding
-                    let mut input = data;
-                    let compressed = encode_szstd_compressed(&mut input);
-                    let decoded = decode_zstd(&compressed).unwrap();
-                    assert_eq!(
-                        decoded, data,
-                        "Decoded data did not match the original input during compression"
-                    );
-                }
-            }
-        }
-    }
+    // The fuzz-artifact interop replay (C-compress -> our-decode and
+    // our-compress -> C-decode) moved to `ffi-bench/tests/fuzz_interop.rs` so
+    // the library crate never links libzstd.
 
     /// Homogeneous input — every byte the same — must NOT be split:
     /// both border histograms are identical (all 512 hits on a single
